read only

+++ internal/sqltest/local/postgres.go
@@ -3,21 +3,21 @@ package local
 import (
 	"context"
 	"fmt"
+	"hash/fnv"
 	"net/url"
 	"os"
 	"strings"
-	"sync"
 	"testing"
 
 	"github.com/jackc/pgx/v5"
-	"github.com/jackc/pgx/v5/pgxpool"
+	"golang.org/x/sync/singleflight"
 
 	migrate "github.com/sqlc-dev/sqlc/internal/migrations"
+	"github.com/sqlc-dev/sqlc/internal/pgx/poolcache"
 	"github.com/sqlc-dev/sqlc/internal/sql/sqlpath"
 )
 
-var postgresPool *pgxpool.Pool
-var postgresSync sync.Once
+var flight singleflight.Group
 
 func PostgreSQL(t *testing.T, migrations []string) string {
 	ctx := context.Background()
@@ -28,16 +28,9 @@ func PostgreSQL(t *testing.T, migrations []string) string {
 		t.Skip("POSTGRESQL_SERVER_URI is empty")
 	}
 
-	postgresSync.Do(func() {
-		pool, err := pgxpool.New(ctx, dburi)
-		if err != nil {
-			t.Fatal(err)
-		}
-		postgresPool = pool
-	})
-
-	if postgresPool == nil {
-		t.Fatalf("PostgreSQL pool creation failed")
+	postgresPool, err := poolcache.New(ctx, dburi)
+	if err != nil {
+		t.Fatalf("PostgreSQL pool creation failed: %s", err)
 	}
 
 	var seed []string
@@ -45,48 +38,71 @@ func PostgreSQL(t *testing.T, migrations []string) string {
 	if err != nil {
 		t.Fatal(err)
 	}
+
+	h := fnv.New64()
 	for _, f := range files {
 		blob, err := os.ReadFile(f)
 		if err != nil {
 			t.Fatal(err)
 		}
+		h.Write(blob)
 		seed = append(seed, migrate.RemoveRollbackStatements(string(blob)))
 	}
+
+	name := fmt.Sprintf("sqlc_test_%x", h.Sum(nil))
 
 	uri, err := url.Parse(dburi)
 	if err != nil {
 		t.Fatal(err)
 	}
-
-	name := fmt.Sprintf("sqlc_test_%s", id())
-
-	if _, err := postgresPool.Exec(ctx, fmt.Sprintf(`CREATE DATABASE "%s"`, name)); err != nil {
-		t.Fatal(err)
-	}
-
 	uri.Path = name
-	dropQuery := fmt.Sprintf(`DROP DATABASE IF EXISTS "%s" WITH (FORCE)`, name)
 
-	t.Cleanup(func() {
-		if _, err := postgresPool.Exec(ctx, dropQuery); err != nil {
-			t.Fatal(err)
+	key := uri.String()
+
+	_, err, _ = flight.Do(key, func() (interface{}, error) {
+		row := postgresPool.QueryRow(ctx,
+			fmt.Sprintf(`SELECT datname FROM pg_database WHERE datname = '%s'`, name))
+
+		var datname string
+		if err := row.Scan(&datname); err == nil {
+			fmt.Println("database already exists", name)
+			// Database already exists
+			return nil, nil
 		}
+
+		fmt.Println("creating database name", name)
+		if _, err := postgresPool.Exec(ctx, fmt.Sprintf(`CREATE DATABASE "%s"`, name)); err != nil {
+			return nil, err
+		}
+
+		dropQuery := fmt.Sprintf(`DROP DATABASE IF EXISTS "%s" WITH (FORCE)`, name)
+
+		cleanup := func() {
+			if _, err := postgresPool.Exec(ctx, dropQuery); err != nil {
+				t.Logf("failed cleaning up: %s", err)
+			}
+		}
+
+		conn, err := pgx.Connect(ctx, uri.String())
+		if err != nil {
+			cleanup()
+			return nil, fmt.Errorf("connect %s: %s", name, err)
+		}
+		defer conn.Close(ctx)
+
+		for _, q := range seed {
+			if len(strings.TrimSpace(q)) == 0 {
+				continue
+			}
+			if _, err := conn.Exec(ctx, q); err != nil {
+				cleanup()
+				return nil, fmt.Errorf("%s: %s", q, err)
+			}
+		}
+		return nil, nil
 	})
-
-	conn, err := pgx.Connect(ctx, uri.String())
 	if err != nil {
-		t.Fatalf("connect %s: %s", name, err)
+		t.Fatalf("create db: %s", err)
 	}
-	defer conn.Close(ctx)
-
-	for _, q := range seed {
-		if len(strings.TrimSpace(q)) == 0 {
-			continue
-		}
-		if _, err := conn.Exec(ctx, q); err != nil {
-			t.Fatalf("%s: %s", q, err)
-		}
-	}
-
-	return uri.String()
+	return key
 }
